+++ src/App.fs
@@ -3,7 +3,7 @@ module App
 open Elmish
 open Elmish.HMR
 
-Program.mkProgram State.init State.update View.rootView
+Router.mkProgramWithNavigation Routes.Articles Routes.Articles Types.NavigateTo Routes.pageParser State.init State.update View.rootView
 |> Program.withReactSynchronous "real-world-fable-app"
 |> Program.withConsoleTrace
 |> Program.run